+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Improved `Makefile`, ditched `tox`, refs #12
   - Non-root Docker user, refs #18
   - Bumped to `zbarcam==2019.907`, refs #20
+  - Handle camera permission gracefull, refs #20
 
 ## [v20190303]